scripts/requirements.txt
weasyprint
pyyaml
